chore: Refactor and update examples
Refactor the most basic examples to be more basic. Also fix an issue with using async_await without the error handler.
