+++ thruster-async-await/src/lib.rs
@@ -6,6 +6,7 @@ use futures_legacy::{Future as FutureLegacy};
 use thruster_core::context::Context;
 use thruster_core::request::{RequestWithParams};
 use thruster_core::route_parser::{MatchedRoute};
+#[cfg(feature = "thruster_error_handling")]
 use thruster_core::errors::Error;
 
 pub fn resolve<R: RequestWithParams, T: 'static + Context + Send>(context_generator: fn(R) -> T, mut request: R, matched_route: MatchedRoute<T>) -> impl FutureLegacy<Item=T::Response, Error=io::Error> + Send {
